pyproject.toml
[project]
name = "claude-kb"
version = "0.2.0"
description = "Universal knowledge base with Qdrant for Claude Code integration"
readme = "README.md"
requires-python = ">=3.13"
authors = [{name = "Artem Khramov", email = "tenequm@gmail.com"}]
license = {text = "MIT"}
keywords = ["knowledge-base", "qdrant", "vector-search", "semantic-search", "claude", "ai", "embeddings"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Database",
    "Topic :: Text Processing :: Indexing",
]
dependencies = [
    "qdrant-client>=1.11.0,<2",
    "sentence-transformers>=5.0.0,<6",
    "torch>=2.0.0,<3",
    "click>=8.1.0,<9",
    "rich>=13.0.0,<15",
    "pydantic>=2.0.0,<3",
    "python-dotenv>=1.0.0,<2",
    "httpx>=0.27.0,<1",
    "tiktoken>=0.5.0,<1",
    "mcp>=1.23.0,<2",
]

[dependency-groups]
dev = [
    "ruff>=0.7.0",
    "ty>=0.0.1a32",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pre-commit>=3.0.0",
    "detect-secrets>=1.5.0",
]

[project.urls]
Homepage = "https://github.com/tenequm/claude-kb"
Repository = "https://github.com/tenequm/claude-kb"
Issues = "https://github.com/tenequm/claude-kb/issues"

[project.scripts]
kb = "claude_kb.cli:main"
claude-kb = "claude_kb.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/claude_kb"]

# =============================================================================
# RUFF: Loose, Helpful Rules (No Christmas Tree!)
# =============================================================================
[tool.ruff]
target-version = "py313"
line-length = 100

# Only check what actually helps catch bugs
[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors (syntax issues that break code)
    "F",   # pyflakes (undefined vars, unused imports - actual bugs!)
    "I",   # isort (import sorting - keeps it clean)
    "UP",  # pyupgrade (use modern Python syntax)
]

# Ignore the annoying stuff that doesn't break code
ignore = [
    "E501",   # Line too long - we have line-length, formatter handles it
    "E741",   # Ambiguous variable name (l, O, I) - sometimes makes sense
    "UP007",  # Use X | Y for unions - Optional[X] is fine and more readable
    "UP006",  # Use type instead of Type - both are valid
]

# Don't check these paths
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    "*.egg-info",
]

[tool.ruff.format]
# Use double quotes
quote-style = "double"

# Indent with spaces
indent-style = "space"

# Unix line endings
line-ending = "lf"

# =============================================================================
# PYTEST
# =============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",  # Show extra test summary info
]

# =============================================================================
# TY: Type Checker (Astral)
# =============================================================================
[tool.ty.environment]
python-version = "3.13"

[tool.ty.src]
include = ["src"]

[tool.ty.rules]
# Keep defaults for most rules - they catch real issues
# Uncomment to enable stricter checks:
# possibly-unresolved-reference = "warn"
